[01-05] ScanResultAdapter.kt: RecyclerView is now ListAdapter of BluetoothConnectableDevice
MainActivity.kt :
Clean up redundant permission things. ListAdapter now scroll to top when the first item changed

DeviceDialogFragment.kt :
Now the content will show when fragment is ready.  It s image of button state and led which can be clicked to turn led on/off

BluetoothService.kt
Scanning creates a device with Bluetooth connectivity
after clicking an item on the Adapter list, the BluetoothConnectableDevice device is selected and the connect (), writeData () and disconnect () functions are run by it